:sparkles: Feat: Refactor invoice handling to use BigDecimal for monetary values; update related calculations and data structures for improved precision and consistency.

+++ src/main/java/com/foryourlife/admin/sales/payments/payment/application/PaymentCommandService.java
@@ -363,11 +363,11 @@ public class PaymentCommandService {
             details.add(new InvoiceContificoJson.Detalle(
                     product.getContificoId(),
                     1,
-                    value.doubleValue(),
+                    value,
                     15,
                     0,
                     0,
-                    value.doubleValue(),
+                    value,
                     0,
                     0,
                     0.0
